Use KineticEnergy instead of getEnergy (#5786)
We currently have two functions in the code that do almost exactly the
same thing:

```
    /**
     * \brief Return (relativistic) particle energy given velocity and mass.
     * Note the use of `double` since this calculation is prone to error with
     * single precision.
     *
     * @param[in] u2 square of particle speed (i.e. u dot u where u = gamma*v)
     * @param[in] mass particle mass
     * @param[out] energy particle energy in eV
     */
    AMREX_GPU_HOST_DEVICE AMREX_INLINE
    void getEnergy ( amrex::ParticleReal const u2, double const mass,
                     double& energy )
    {
        using std::sqrt;
        using namespace amrex::literals;

        constexpr auto c2 = PhysConst::c * PhysConst::c;
        energy = mass * u2 / (sqrt(1.0_rt + u2 / c2) + 1.0_rt) / PhysConst::q_e;
    }
```

and 

```
    AMREX_GPU_HOST_DEVICE AMREX_INLINE
    amrex::ParticleReal KineticEnergy(
        const amrex::ParticleReal ux, const amrex::ParticleReal uy, const amrex::ParticleReal uz,
        const amrex::ParticleReal mass)
    {
        using namespace amrex;

        constexpr auto inv_c2 = 1.0_prt/(PhysConst::c * PhysConst::c);

        // The expression used is derived by reducing the expression
        // (gamma - 1)*(gamma + 1)/(gamma + 1)

        const auto u2 = ux*ux + uy*uy + uz*uz;
        const auto gamma = std::sqrt(1.0_prt + u2*inv_c2);
        return 1.0_prt/(1.0_prt + gamma)*mass*u2;
    }
```
The only differences are that `getEnergy` forces the use of double
precision and returns an energy in `eV`. I think that it would be better
if we could use only one function for clarity. Therefore, in this PR I
replaced `getEnergy` with `KineticEnergy`.

The units of the output are easy to deal with (we can just divide by 1eV
the result). The precision is trickier. @roelof-groenewald , do you
think that we really need to force `double` precision here? If it is the
case, I think that it could be preferable to make `getEnergy` a template
function and force double precision with something like
`KineticEnergy<double>(...)`. What do you think?

---------