CI: Fix Skipped Required Workflows (again) (#5907)
Follow up on #5889 and fix the issue that appeared in #5883, related to
the fact that we add the `clang-tidy` checks automatically from a
matrix, rather than manually as separate jobs in the workflow file.

Note that I had managed to fix a similar issue in the alternative
in-house implementation I had proposed in #5469.

To-do:
- [x] Fix main issue by applying the `if` condition at the `step` level
(rather than the `job` level)
- [x] Remove `cron` schedule for CodeQL workflows (on Sundays at 3:27
AM...?)
- [x] Revert change in .github/workflows/check_changes.yml (implemented
only to test the PR)

Future PRs:
- Skip also CodeQL workflows (C++ and Python files analysis) - not
working yet...