Add feature running variance deposition (#5811)
is a single pass algorithm that can compute current and variance using
accumulation buffers to mitigate requiring a second pass through the
particle list when computing variance. It is used to get temperature
deposition with matching shape functions as the current density
multifabs for applying relaxation algorithms in the hybrid solver that
depend on computing ion moments.

---------

Signed-off-by: S. Eric Clark <25495882+clarkse@users.noreply.github.com>