`PlasmaInjector.H`: Hide `openPMD.hpp`
Avoid the `openPMD/openPMD.hpp` include in `PlasmaInjector.H`,
because the latter file is included in tons of translation
units.

Use a PIMPL-like idom with `std::any` to still store the series
as a member variable, but only pull in the type in the two
specific translation units, `AddParticles.cpp` and `PlasmaInjector.cpp`
that access the member.

All other openPMD includes look good an TU local.