Load density from openPMD file
This adds a new capability of loading density from an openPMD file. It works
for initialization and continuous injection. It supports 3d, 2d rz, 2d xz
and 1d z coordinates. The openPMD data must have the C order for
multi-dimensional arrays. This restriction could be relaxed in the
future. Currently, it loads all the data from the file. In future, we could
try to load data by chunks.