Fix bug for field probe + moving window
FieldProbe only probes data while the whole probe is inside of the simulation space. In the existing implementation, it does that check by comparing the configured settings (e.g. z_probe) vs simulation space. For moving window cases, the initial configured setting does not reflect the probe point. This can lead to two issues:
1. Incorrectly determining that the probes are outside of the simulation space, and not generating data (issue 6082)
2. Incorrectly determining that the probes are inside of the simulation space
The second issue can happen in two ways: if the geometry requested goes outside (e.g. x1_probe for a Line is outside of the simulation space), or if the sliding window causes the probe location to be outside of the simulation space, but to include the initial configured location. The second issue may result in a memory access violation.

Fix the issue by checking the actual location of the probes against the simulation space -- after applying moving window -- instead of checking the initial configuration settings of the FieldProbe.