Load Density: Distributed Approach
Previously, the data are duplicated on all processes. For a very large file,
this might be problematic. A distributed approach is developed in this PR.

For the initial particle creation, we load the data we need in a distributed
way. Because we also know the particle BoxArray and DistributionMapping, we
can then redistribute the data to where they are needed.

For the continuously particle injection during moving window, the data are
duplicated on every process, but they are still distributed in time in the
sense that we only load the data needed for the new region created by the
moving window plus some extra caching so that we don't have to do disk I/O
too frequently.

A caveat is that it might be difficult to determine the range of the data we
need if the boosted frame is used and the boosted velocity is not a
constant. In that case, one might encounter an out-of-bound error at run
time. One can also disable the distribute approach with a run time
parameter, `TBD`.