Remove WarpX::GetInstance
We now pass a `WarpX*` to where we previously needed
`WarpX::GetInstance()`. For classes, this means storing a pointer to `WarpX`
as a member variable either in the class itself or in its base class. We use
`WarpX*` rather than `WarpX const*` as the data type to avoid frequent use
of `const_cast` in practice.