Checkpoint/Plotfile: Robust Bounds Calc
The calculation of the bounds in plotfiles (incl. checkpoints)
appears fishy. It relies on the order of the boxes in the
box array, which is not always pointing to a box on the
lower/upper bound of the simulation when load balancing is
used.

This can corrupt checkpoints (plotfiles) when writing, by
a) filtering out valid particles & fields and b) remembering
the wrong simulation meta-data for the simulation geometry
of fields/particles.

Co-authored-by: Andrew Myers <atmyers@lbl.gov>
Co-authored-by: Remi Lehe <rlehe@lbl.gov>
Co-authored-by: Davide Terzani <dterzani@lbl.gov>