Custom weights for initial DistributionMapping
By default, amrex uses box volumes as weights when distributing boxes. But
this would somewhat defeat the purpose of splitting high density boxes until
the next load balance is performed. Thus, when splitting box is on, we are
going to assume that the boxes have equal weights.

For a test based on #6158, this improves the initial load balance efficiency
from 0.26 to 0.63 when running on 8 processes.