Custom weights for initial DistributionMapping (#6452)
By default, amrex uses box volumes as weights when distributing boxes.
But this would somewhat defeat the purpose of splitting high density
boxes until the next load balance is performed. Thus, when splitting box
is on, we are going to assume that the boxes have equal weights.

For a test based on #6158, this improves the initial load balance
efficiency from 0.26 to 0.63 when running on 8 processes.

The figures below show the distribution mapping without and with this
PR.

<img width="844" height="636" alt="Screenshot from 2025-12-11 16-07-56"
src="https://github.com/user-attachments/assets/eb6b4c85-fc55-4e4e-ac68-b6a02ab9bf4d"
/>