External Particle Fields: enable reading of multiple fields from file (#6269)
This PR adds the functionality to read in multiple external particle
fields from an openPMD file. Inspired from the HybridSolver, a new
helper class `ExternalParticleFields` is introduced, which holds the
arrays of the meta data for the external E and B fields.

Then, the fields are read and written into the `m_fields` by using as
many components as there are external E and B fields, respectively. This
allows to read in fields from different files witth different time
dependencies.

The previous syntax of reading in a single field is preserved and
functional. Now, additionally, multiple fields can be read in as
```
            particles.B_ext_particle_init_style = read_from_file
            particles.B_ext_particle_fields = b1 b2
            particles.b1.read_fields_from_path = diags/Bfield_map1
            particles.b1.read_fields_B_dependency(t) = cos(omega*t + phase)
            particles.b2.read_fields_from_path = diags/Bfield_map2
            particles.b2.read_fields_B_dependency(t) = cos(2*omega*t + phase)
```

For the picmi interface, a dictionary has to be created per field,
similarly to the HybridSolver:
```python
    applied_field = picmi.LoadAppliedField(
        load_E=False,
        load_B=True,
        B_external_fields={
            "b1": {
                "read_fields_from_path": "diags/Bfield_map1",
                "read_fields_B_dependency(t)": f"cos({omega}*t + {phase})",
            },
            "b2": {
                "read_fields_from_path": "diags/Bfield_map2",
                "read_fields_B_dependency(t)": f"cos(2*{omega}*t + {phase})",
            },
        },
    )
```


Note that the global `particles.read_fields_from_path` is ignored as
soon as multiple fields are specified and a warning is issued.
Furthermore, for each field, the path must be specified individually.

Additional CI tests are added with different time dependency, similar to
the previously existing one that read only a single field.

This plot shows the beating from two field maps at different
frequencies:

![frequency_beating](https://github.com/user-attachments/assets/d7fb8353-76bd-4789-bcd0-85210adaac9d)

Any feedback or suggestions are more than welcome!

---------

Co-authored-by: Edoardo Zoni <59625522+EZoni@users.noreply.github.com>