collisions should be after call to HandleParticlesAtBoundaries() (#6458)
This PR moves the collisions for the implicit solver to the proper place
in the Evolve loop. This fixes a bug introduced in PR #6428, where the
collisions are moved from just before to just after the PIC advance.
(See https://github.com/BLAST-WarpX/warpx/issues/6449). However, with
this change I noticed odd behavior in simulation outputs at processor
boundaries late in time for highly collisional simulations. The fix is
that the collisions should be placed after the particle communication
and boundary handling is performed, which occurs later in the Evolve
loop.