check for out of bounds particles with cropping in suborbits. (#6467)
Particle suborbits were added in PR #5969. Suborbits are needed for
improved accuracy of particle orbits when using large time steps.
Particle orbit cropping at physical boundaries was added in PR #5649.
This is used to achieve rigorous charge conservation for outflow
particles at physical boundaries (conductor and insulator).

A bug occurs when the initial position of the particle is out of bounds
(past the physical boundary) where cropping is being done. In this
scenario, the particle will deposit current when it should not be doing
so, breaking energy and charge conservation. This PR fixes this bug by
checking for particles with the initial position out of bounds in the
suborbit routine and neglecting deposition for these suborbits.